Completed PHPUnit setup. Fixed type in utility function 'formatNumber'. Added second testcase for utility functions. Removed test index script.

+++ backend/src/php/de/mayflower/nada/util/Util.php
@@ -10,11 +10,11 @@
         /**
          *  Formats the given number.
          *
-         *  @param int $number The number to format.
+         *  @param float $number The number to format.
          *
          *  @return string The formatted number with two decimal digits.
          */
-        public final function formatNumber( int $number ) : string
+        public final function formatNumber( float $number ) : string
         {
             return number_format( $number, 2 );
         }
